Fixing bug with material temp group
To make sure a material exists in the model before painting it onto a buidling it's drawn to a temporary group inside the buildign group. This group was previosly removed when the buidling was redrawn without being replaced which resluted in material replacement not working in buidling properties once the user had clicked Apply.